added support for finding geocaches
user can scan qr code and mark the geocache as found and saves the time in database. In SearchScreen after geocache is found a marker will appear. GoogleMap component loads all hidden or found markers depending on the current screen. Also added QRScanButton for simpler reusable option to navigate to the QRScanScreen.